// src/app/modules/ssid/models/sssid.models.ts
export class SsidModel {
    realName: string;
    dob: string;
    gender: string;
    documentID: string;
    documentType: string;
    documentImage: string | ArrayBuffer | null;
    userImageWithOTP: string | ArrayBuffer | null;
    streetAddress: string;
    country: string;
    state: string;
    city: string;
  }